+++ cryptolyzer/ike/curves.py
@@ -11,8 +11,9 @@ from cryptodatahub.ike.algorithm import (
 
 from cryptodatahub.ike.version import IkeVersion
 
+from cryptolyzer.common.result import AnalyzerResultIKE, AnalyzerTargetIke
+
 from cryptolyzer.ike.dh import AnalyzerDHBase
-from cryptolyzer.common.result import AnalyzerResultIKE
 
 
 @attr.s
@@ -66,7 +67,7 @@ class AnalyzerCurves(AnalyzerDHBase):
         dh_groups, key_reused = self._analyze(analyzable, protocol_version)
 
         return AnalyzerResultCurves(
-            target=analyzable,
+            target=AnalyzerTargetIke.from_l7_client(analyzable),
             groups=dh_groups,
             key_reused=key_reused,
         )