+++ CHANGELOG.rst
@@ -2,6 +2,19 @@
 Changelog
 =========
 
+----------
+Unreleased
+----------
+
+Features
+========
+
+-  IKE (``ike``)
+
+   -  Extensions (``extensions``)
+
+      -  add checker extensions checker (#177)
+
 ------------------
 1.3.0 - 2026-06-15
 ------------------